app/requirements.app.txt
Flask==3.1.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
marshmallow==3.23.2
gunicorn==25.1.0